fix: lower BIN_MAX_FILES to 3 on measured recall data, add grammar-rejection retry rung for Gemini 400s
