+++ src/server/core/review/pack.ts
@@ -10,7 +10,19 @@ export const PACKABLE_MAX_DIFF_LINES = 150;
 // larger bin "saved" get spent back, and every file in that bin waits out the long call first.
 export const BIN_TARGET_DIFF_LINES = 300;
 // Blast radius and attention: per-file recall degrades before the token budget runs out.
-export const BIN_MAX_FILES = 6;
+//
+// Lowered from 6 to 3 on measurement, not intuition. Benchmarked over 787 reviews of a 12-file diff
+// carrying 35 planted defects, holding everything else fixed and overriding only this grouping, recall
+// fell monotonically as the bin grew: 53.6% at two files per call, 46.3% at three, 41.5% at six, 34.3%
+// at twelve. Precision was 100% at every size, so the loss is purely defects never mentioned. The
+// responses stayed well inside their token budget throughout -- what degrades is attention across
+// files, which is why no amount of extra output room recovers it.
+//
+// The cost is real and paid in subrequests, not tokens per finding: halving the bin doubles the number
+// of model calls per job (a 221-file job goes from ~37 bins to ~74), and each call re-sends the
+// ~2,800-token preamble. On the Workers Free 50-subrequests-per-invocation ceiling that means more
+// continuations per job, which phase-control already handles but which shows up as wall clock.
+export const BIN_MAX_FILES = 3;
 export const BIN_DIFF_CHAR_BUDGET = 24_000;
 
 export type ReviewUnit =
